refactor(backend): add custom time for tokens
