+++ backend/Zeus.Api.Domain/Authentication/ValueObjects/RefreshToken.cs
@@ -1,3 +1,5 @@
+using Zeus.Api.Domain.Authentication.ValueObjects;
+
 namespace Zeus.Api.Domain.Common.ValueObjects.Authentication;
 
 public sealed class RefreshToken : Token
