+++ backend/Zeus.Api.Infrastructure/Authentication/JwtSettings.cs
@@ -5,7 +5,8 @@ public class JwtSettings
     public const string SectionName = "JwtSettings";
     
     public string Secret { get; init; } = null!;
-    public int ExpiryMinutes { get; init; }
+    public int AccessTokenExpiryMinutes { get; init; }
+    public int RefreshTokenExpiryMinutes { get; init; }
     public string Issuer { get; init; } = null!;
     public string Audience { get; init; } = null!;
 }
