+++ backend/Zeus.Api.Web/appsettings.Development.json
@@ -7,7 +7,8 @@
   },
   "JwtSettings": {
     "Secret": "",
-    "ExpiryMinutes": 60,
+    "AccessTokenExpiryMinutes": 60,
+    "RefreshTokenExpiryMinutes": 43200,
     "Issuer": "Triggo Development Api",
     "Audience": "Triggo Developer"
   }
