+++ backend/Zeus.Api.Web/appsettings.json
@@ -8,7 +8,8 @@
   "AllowedHosts": "*",
   "JwtSettings": {
     "Secret": "",
-    "ExpiryMinutes": 60,
+    "AccessTokenExpiryMinutes": 60,
+    "RefreshTokenExpiryMinutes": 43200,
     "Issuer": "Triggo Api",
     "Audience": "Triggo Client"
   }
